MB-69666: Fix -Werror,-Wimplicit-int-conversion
When building with Clang 18 on Ubuntu 24 there is various
errors reported caused by passing size_t to htons()

Reviewed-on: https://review.couchbase.org/c/kv_engine/+/237286
Tested-by: Build Bot <build@couchbase.com>

+++ engines/ep/src/getkeys.cc
@@ -86,7 +86,7 @@ void AllKeysCallback::callback(const DiskDocKey& key) {
         }
     }
 
-    uint16_t outlen = htons(outKey.size());
+    uint16_t outlen = htons(gsl::narrow_cast<uint16_t>(outKey.size()));
     // insert 1 x u16
     const auto* outlenPtr = reinterpret_cast<const char*>(&outlen);
     buffer.insert(buffer.end(), outlenPtr, outlenPtr + sizeof(uint16_t));
